Refactor visit files
