+++ integration_tests/pages/visit/cancel/cancelVisitPage.ts
@@ -1,5 +1,5 @@
 import { type Locator, type Page } from '@playwright/test'
-import AbstractPage from '../abstractPage'
+import AbstractPage from '../../abstractPage'
 
 export default class CancelVisitPage extends AbstractPage {
   readonly axeExcludedElements = [
