+++ integration_tests/pages/visit/cancel/visitCancelledPage.ts
@@ -1,5 +1,5 @@
 import { type Locator, type Page } from '@playwright/test'
-import AbstractPage from '../abstractPage'
+import AbstractPage from '../../abstractPage'
 
 export default class VisitCancelledPage extends AbstractPage {
   readonly visitDetails: Locator
