+++ integration_tests/pages/visit/clearNotifications/clearNotificationsPage.ts
@@ -1,5 +1,5 @@
 import { type Locator, type Page } from '@playwright/test'
-import AbstractPage from '../abstractPage'
+import AbstractPage from '../../abstractPage'
 
 export default class ClearNotificationsPage extends AbstractPage {
   // Known issue with radio conditional reveal
