+++ integration_tests/pages/visit/details/visitDetailsPage.ts
@@ -1,5 +1,5 @@
 import { type Locator, type Page } from '@playwright/test'
-import AbstractPage from '../abstractPage'
+import AbstractPage from '../../abstractPage'
 
 export default class VisitDetailsPage extends AbstractPage {
   // Visit details
