+++ integration_tests/specs/request/processVisitRequest.spec.ts
@@ -2,7 +2,7 @@ import { test, expect } from '@playwright/test'
 import TestData from '../../../server/routes/testutils/testData'
 import HomePage from '../../pages/homePage'
 import VisitRequestsListingPage from '../../pages/request/visitRequestsListingPage'
-import VisitDetailsPage from '../../pages/visit/visitDetailsPage'
+import VisitDetailsPage from '../../pages/visit/details/visitDetailsPage'
 import orchestrationApi from '../../mockApis/orchestration'
 import { resetStubs, login } from '../../testUtils'
