+++ integration_tests/specs/review/reviewAVisit.spec.ts
@@ -1,8 +1,8 @@
 import { test, expect } from '@playwright/test'
 import { format, add } from 'date-fns'
 import TestData from '../../../server/routes/testutils/testData'
-import VisitDetailsPage from '../../pages/visit/visitDetailsPage'
-import ClearNotificationsPage from '../../pages/visit/clearNotificationsPage'
+import VisitDetailsPage from '../../pages/visit/details/visitDetailsPage'
+import ClearNotificationsPage from '../../pages/visit/clearNotifications/clearNotificationsPage'
 import eventAuditTypes from '../../../server/constants/eventAudit'
 import { notificationTypeAlerts, notificationTypes } from '../../../server/constants/notifications'
 import orchestrationApi from '../../mockApis/orchestration'
