+++ integration_tests/specs/visit/cancelVisit.spec.ts
@@ -1,9 +1,9 @@
 import { test, expect } from '@playwright/test'
 import { format, add } from 'date-fns'
 import TestData from '../../../server/routes/testutils/testData'
-import VisitDetailsPage from '../../pages/visit/visitDetailsPage'
-import CancelVisitPage from '../../pages/visit/cancelVisitPage'
-import VisitCancelledPage from '../../pages/visit/visitCancelledPage'
+import VisitDetailsPage from '../../pages/visit/details/visitDetailsPage'
+import CancelVisitPage from '../../pages/visit/cancel/cancelVisitPage'
+import VisitCancelledPage from '../../pages/visit/cancel/visitCancelledPage'
 import HomePage from '../../pages/homePage'
 import { CancelVisitOrchestrationDto } from '../../../server/data/orchestrationApiTypes'
 import orchestrationApi from '../../mockApis/orchestration'
