+++ integration_tests/specs/visitJourney/updateAVisit.spec.ts
@@ -1,9 +1,9 @@
 import { test, expect } from '@playwright/test'
 import { format, sub, addDays, eachDayOfInterval } from 'date-fns'
 import TestData from '../../../server/routes/testutils/testData'
-import VisitDetailsPage from '../../pages/visit/visitDetailsPage'
+import VisitDetailsPage from '../../pages/visit/details/visitDetailsPage'
 import SelectVisitorsPage from '../../pages/visitJourney/selectVisitorsPage'
-import ConfirmUpdatePage from '../../pages/visit/confirmUpdatePage'
+import ConfirmUpdatePage from '../../pages/visit/update/confirmUpdatePage'
 import SelectVisitDateAndTimePage from '../../pages/visitJourney/selectVisitDateAndTimePage'
 import AdditionalSupportPage from '../../pages/visitJourney/additionalSupportPage'
 import MainContactPage from '../../pages/visitJourney/mainContactPage'
