+++ server/routes/visit/index.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express'
 import { BadRequest } from 'http-errors'
 import { Services } from '../../services'
-import VisitDetailsController from './visitDetailsController'
-import CancelVisitController from './cancelVisitController'
+import VisitDetailsController from './details/visitDetailsController'
+import CancelVisitController from './cancel/cancelVisitController'
 import { isValidVisitReference } from '../validationChecks'
-import ClearNotificationsController from './clearNotificationsController'
-import ConfirmUpdateController from './confirmUpdateController'
-import UpdateVisitController from './updateVisitController'
-import ProcessVisitRequestController from './processVisitRequestController'
+import ClearNotificationsController from './clearNotifications/clearNotificationsController'
+import ConfirmUpdateController from './update/confirmUpdateController'
+import UpdateVisitController from './update/updateVisitController'
+import ProcessVisitRequestController from './visitRequests/processVisitRequestController'
 import VisitPassesController from '../visitPasses/visitPassesController'
 
 export default function routes(services: Services): Router {
