+++ server/routes/visit/update/confirmUpdateController.ts
@@ -1,8 +1,8 @@
 import { RequestHandler, Request } from 'express'
 import { BadRequest } from 'http-errors'
-import { isValidVisitReference } from '../validationChecks'
-import { VisitReferenceParams } from '../../@types/requestParameterTypes'
-import { appendNavStateToPath, extractVisitNavState } from './visitNavigationUtils'
+import { isValidVisitReference } from '../../validationChecks'
+import { VisitReferenceParams } from '../../../@types/requestParameterTypes'
+import { appendNavStateToPath, extractVisitNavState } from '../visitNavigationUtils'
 
 export default class ConfirmUpdateController {
   public constructor() {}
@@ -13,7 +13,7 @@ export default class ConfirmUpdateController {
       const { policyNoticeDaysMin } = req.session.selectedEstablishment
       const navState = extractVisitNavState({ from: req.query.from, query: req.query.query })
 
-      return res.render('pages/visit/confirmUpdate', {
+      return res.render('pages/visit/update/confirmUpdate', {
         errors: req.flash('errors'),
         backLinkHref: appendNavStateToPath(`/visit/${reference}`, navState),
         formAction: appendNavStateToPath(`/visit/${reference}/confirm-update`, navState),
