+++ server/routes/visit/update/updateVisitController.ts
@@ -1,12 +1,12 @@
 import { RequestHandler } from 'express'
 import { differenceInCalendarDays, format, parseISO } from 'date-fns'
-import { VisitService } from '../../services'
-import { VisitReferenceParams } from '../../@types/requestParameterTypes'
-import { clearSession } from '../visitorUtils'
-import { VisitSessionData } from '../../@types/bapv'
-import { convertToTitleCase } from '../../utils/utils'
-import { getIdsToFlag, getPrisonerLocation, isPublicBooking } from './visitUtils'
-import { appendNavStateToPath, extractVisitNavState } from './visitNavigationUtils'
+import { VisitService } from '../../../services'
+import { VisitReferenceParams } from '../../../@types/requestParameterTypes'
+import { clearSession } from '../../visitorUtils'
+import { VisitSessionData } from '../../../@types/bapv'
+import { convertToTitleCase } from '../../../utils/utils'
+import { getIdsToFlag, getPrisonerLocation, isPublicBooking } from '../visitUtils'
+import { appendNavStateToPath, extractVisitNavState } from '../visitNavigationUtils'
 
 export default class UpdateVisitController {
   public constructor(private readonly visitService: VisitService) {}
